third_party/WebKit/LayoutTests/fast/selectors/style-sharing-children-prevent-sharing.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
.root + #lastChildRules {
    color: red;
}
</style>

<div id="y" class="root">
    <div></div>
</div>

<div id="x" class="root">
    <div></div>
</div>

<script>
description("Test dynamic changes to the childrenSupportStyleSharing flag.");

var x = document.getElementById("x");
var y = document.getElementById("y");

// Attach the whole tree, this makes x and y share, and the children of x and y share.
document.body.offsetTop;

// Add a child that make us match sibling rules. This will set the ChildrenAffectedByDirectAdjacentRules
// flag on #x preventing its children from sharing in the future.
var lastChildRules = x.appendChild(document.createElement("div"));
lastChildRules.id = "lastChildRules";
x.appendChild(document.createElement("div"));

// Add a new child to #y. Normally it could share with children of #x since both #x and #y share,
// but now #x has children affected by RestyleFlags so they can no longer share.
// FIXME: This element could technically share with the other <div> inside #y, but recalcStyle
// goes from lastChild -> firstChild so we haven't added the firstChild (which doesn't need a recalc)
// to the candidate list yet.
y.appendChild(document.createElement("div"));

document.body.offsetTop;

if (window.internals) {
    shouldBeTrue("internals.isSharingStyle(x, y)");
    shouldBeTrue("internals.isSharingStyle(x.firstElementChild, y.firstElementChild)");
    shouldBeFalse("internals.isSharingStyle(x.firstElementChild, y.firstElementChild.nextElementSibling)");
    shouldBeFalse("internals.isSharingStyle(lastChildRules, x.firstElementChild)");
    shouldBeFalse("internals.isSharingStyle(lastChildRules, y.firstElementChild)");

    // FIXME: We should see if we can look at direct siblings somehow since these elements do
    // match the same rules at the same depth, recalcStyle is just in the wrong order to allow
    // them to share.
    shouldBeFalse("internals.isSharingStyle(y.firstElementChild, y.firstElementChild.nextElementSibling)");
}
</script>
