third_party/WebKit/LayoutTests/fast/selectors/style-sharing-children-prevent-sharing-expected.txt

Test dynamic changes to the childrenSupportStyleSharing flag.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isSharingStyle(x, y) is true
PASS internals.isSharingStyle(x.firstElementChild, y.firstElementChild) is true
PASS internals.isSharingStyle(x.firstElementChild, y.firstElementChild.nextElementSibling) is false
PASS internals.isSharingStyle(lastChildRules, x.firstElementChild) is false
PASS internals.isSharingStyle(lastChildRules, y.firstElementChild) is false
PASS internals.isSharingStyle(y.firstElementChild, y.firstElementChild.nextElementSibling) is false
PASS successfullyParsed is true

TEST COMPLETE
